refactor: simplify presence/preferences wiring
- drop TOCTOU existsSync in readPreferencesFile; catch ENOENT
- collapse broadcastCursorMotion/broadcastCursorSplineViz into helper
- schedulePresenceExpiry: skip reschedule if timer already pending
- extract FilledCursorIcon to renderer/shared so debug bundle
  no longer imports from canvas-bg
